Wizzers/Chat-Bot/simple_requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
